hub/pyproject.toml
[tool.poetry]
name = "sonoff-edge"
version = "0.1.0"
description = ""
authors = ["mystifier <mystifier0505@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.5"
pycryptodome = "^3.18.0"
zeroconf = "^0.103.0"
netifaces = "^0.11.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
